Remove Ledge Intangibility Scaling
